Updated to use mapbox hpp skel, made into header only library, added google benchmark, fixed various configuration issues, moved the namespace some
