+++ test/test-files/osm-nodes-45331-triangles.json
